Clean project structur && norm

+++ project/src/pars_check.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   pars_check.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jmartin <jmartin@student.42.fr>            +#+  +:+       +#+        */
+/*   By: jmartin <jmartin@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/28 16:09:15 by jmartin           #+#    #+#             */
-/*   Updated: 2022/08/29 11:12:38 by jmartin          ###   ########.fr       */
+/*   Updated: 2022/08/30 01:55:16 by jmartin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,10 +38,8 @@ void	check_player_pos(t_game *game)
 				|| game->map->scene[i][j] == 'E'
 				|| game->map->scene[i][j] == 'W')
 			{
-				game->player->pos_y = ((double)(i * TILE_SIZE))
-					+ TILE_SIZE / 2 + MAPOS;
-				game->player->pos_x = ((double)(j * TILE_SIZE))
-					+ TILE_SIZE / 2 + MAPOS;
+				game->player->pos_y = (i * TILE_SIZE) + TILE_SIZE / 2 + MAPOS;
+				game->player->pos_x = (j * TILE_SIZE) + TILE_SIZE / 2 + MAPOS;
 				game->player->skin = game->map->scene[i][j];
 				game->player->is_ready = 1;
 				break ;
